+++ ts/test/session/integration/closed_group_itest.ts
