+++ ts/test/session/integration/registration_itest.ts
